Ignore rows with empty title
Kairos throws a fatal exception when encountering an element / row that has no title.

The reason for the crash is that the current code assumes that when there's no title the element must be a checkbox. That's not the case. It can also be just an element with no title.

In my testing reading the title of checkbox elements didn't require any special lookup. The element.text just works. That's why I've removed that code from the else statement and replaced it with a warning.

The consequence of this change is that elements which have no title will not appear in default_inputs.

Example pine code to test reading of checkbox title using element.text
checkbox = input.bool(true, title='my checkbox')

Example pine code to test ignoring of rows with no title.
input.color(title='', defval=color.new(#D1D4DC, 0), inline='FILL_COLORS_INLINE_1')