Add check for strategies with 0 trades
Since nov 2022 TV has disabled the export list button on strategies that have 0 trades. In the past the button would be enabled and just download an empty csv file. To cope with this new situation we have to do two things differently. First we need a different way to determine the default filename as we can't reliably extract that from the first download anymore. Second we need to check if the strategy has trades before we attempt to click the export button.