Switch to new headless mode
I ran into a problem with Chrome driver version 110 where it would throw an `unable to discover open pages` error when using the shared folder setting on a headless machine. Switching to the new headless mode fixes the issue.  

This is a breaking change as it will require using a Chrome driver version 110+  so proceed with care.